Integration test for configure OTE
Integration testing to check whether configure OTE successfully updates the SAP Agent configuration file and whether the changes get reflected in the behaviour of the agent.

We verify the collection of 4 features: Process Metrics, Agent Metrics, HANA Monitoring and Workload Validation.
We check for enabling and disabling of all the metrics and see if the metrics get logged to cloudmonitoring. For process metrics, we also check if certain metrics are successfully skipped for collection when added to 'process_metrics_to_skip' field.

PiperOrigin-RevId: 595300007